django_backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=2.9
django-cors-headers>=4.0
